requirements.txt
flask
flask_session
Flask-SQLAlchemy
pymysql
requests